services/requirements.txt
SQLAlchemy==1.4.36
psycopg2-binary==2.9.3
python-dotenv==0.20.0
Flask==2.1.2
Flask-Login==0.6.1
loguru==0.6.0
py-cord==2.0.0b7